+++ src/main/java/Amogus/AmogusException.java
@@ -1,3 +1,5 @@
+package Amogus;
+
 public class AmogusException extends Exception {
 
     public AmogusException(String message) {
