+++ src/main/java/Amogus/UI.java
@@ -1,3 +1,7 @@
+package Amogus;
+
+import Tasks.TaskList;
+
 import java.util.Scanner;
 
 public class UI {
