+++ src/main/java/Command/MarkCommand.java
@@ -1,13 +1,19 @@
-public class DeleteCommand implements Command {
+package Command;
+
+import Amogus.FileStorage;
+import Amogus.UI;
+import Tasks.TaskList;
+
+public class MarkCommand implements Command {
 
     private int idx;
 
-    public DeleteCommand(int idx) {
+    public MarkCommand(int idx) {
         this.idx = idx;
     }
 
     /**
-     * Removes the specified task using its index.
+     * Marks the task as done.
      *
      * @param tasks List of tasks
      * @param ui any display back to the user
@@ -15,8 +21,8 @@ public class DeleteCommand implements Command {
      */
     @Override
     public void execute(TaskList tasks, UI ui, FileStorage f) {
-        ui.showMsg(ui.format("Noted. I've removed this task:\n  " + tasks.taskDesc(idx) + "\n"));
-        tasks.delete(idx);
+        tasks.mark(idx);
+        ui.showMsg(ui.format("Nice! I've marked this task as done:\n  " + tasks.taskDesc(idx) + "\n"));
         f.saveTasks(tasks);
     }
