+++ src/main/java/Command/UnmarkCommand.java
@@ -1,13 +1,19 @@
-public class MarkCommand implements Command {
+package Command;
+
+import Amogus.FileStorage;
+import Amogus.UI;
+import Tasks.TaskList;
+
+public class UnmarkCommand implements Command {
 
     private int idx;
 
-    public MarkCommand(int idx) {
+    public UnmarkCommand(int idx) {
         this.idx = idx;
     }
 
     /**
-     * Marks the task as done.
+     * Marks the task as not done.
      *
      * @param tasks List of tasks
      * @param ui any display back to the user
@@ -15,8 +21,8 @@ public class MarkCommand implements Command {
      */
     @Override
     public void execute(TaskList tasks, UI ui, FileStorage f) {
-        tasks.mark(idx);
-        ui.showMsg(ui.format("Nice! I've marked this task as done:\n  " + tasks.taskDesc(idx) + "\n"));
+        tasks.unmark(idx);
+        ui.showMsg(ui.format("OK, I've marked this task as not done yet:\n  " + tasks.taskDesc(idx) + "\n"));
         f.saveTasks(tasks);
     }
